% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_period_variation.R
\name{calculate_period_variation}
\alias{calculate_period_variation}
\title{Calculate the exchange rate variation between two dates}
\usage{
calculate_period_variation(
  basedados = dolar,
  DataInicial = "01/01/2022",
  DataFinal = Sys.Date()
)
}
\arguments{
\item{basedados}{A data frame with the exchange rate data, in which the first column is named "Data" and the second column is named "valor".}

\item{DataInicial}{A character string or a date, representing the start date of the period to be analyzed.}

\item{DataFinal}{A character string or a date, representing the end date of the period to be analyzed.}
}
\value{
A data frame with the period between the two dates, the initial and final dates, the initial and final exchange rate, the number of days in the period, the relative variation in the exchange rate, and the absolute variation.
}
\description{
This function receives the exchange rate data and two dates, calculates the exchange rate at the closest date before each of the input dates, and returns a data frame with the period between the two dates, the initial and final dates, the initial and final exchange rate, the number of days in the period, the relative variation in the exchange rate, and the absolute variation.
}
\examples{
\dontrun{
df <- dados_dolar()
datas <-
tibble(
DataInicial = lubridate::dmy("31/10/2022") \%m+\% months(seq(from = 1, to = 12, by = 1)),
DataFinal = lubridate::dmy("30/11/2022") \%m+\% months(seq(from = 1, to = 12, by = 1))
)
tabela <- pmap_dfr(datas,variacao_dolar)
}

}
