% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correct_inflation.R
\name{correct_inflation}
\alias{correct_inflation}
\title{Reajusta um valor de acordo com a inflação}
\usage{
correct_inflation(valor = 1000, database = "01/2001", indice = "ipca")
}
\arguments{
\item{valor}{Valor a ser reajustado (default é 1000.00).}

\item{database}{Data base para o reajuste (default é "01/2020").}

\item{indice}{Indice de inflação a ser utilizado (default é "igpm"). Os índices disponíveis sã: ipca, igpm, igpdi, ipc, inpc

\code{ipca}, \code{igpm},\code{igpdi},\code{ipc},\code{incc}, and \code{inpc}.
Foi utilizado o INCC-DI, para o reajuste pelo INCC.}
}
\value{
A \code{tibble}, com os valor corrigido, e os índices.
}
\description{
\code{correct_inflation()} usa dados do IPEADATA para reajustar um valor, de acordo com um índice.
}
\examples{
# Exemplo, corrigindo pelo ipca:
# correct_inflation(valor = 1000, database = "01/2022", indice = "ipca")

}
\references{
Os índeces podem ser acessados através do acesso ao sítio do IPEA Data
\url{http://www.ipeadata.gov.br/}.
}
