% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correct_interest.R
\name{correct_interest}
\alias{correct_interest}
\title{Correção de um valor pelo cdi ou pela selic}
\usage{
correct_interest(
  valor = 1000,
  dtInicio = "01/01/2003",
  dtFim = Sys.Date(),
  indice = "cdi"
)
}
\arguments{
\item{valor}{Valor a ser reajustado (default é 1000.00).}

\item{dtInicio}{Data de início da atualização.}

\item{dtFim}{Data final da atualização.}

\item{indice}{Indice de inflação a ser utilizado (default é "igpm"). Os índices disponíveis sã: ipca, igpm, igpdi, ipc, inpc

\code{ipca}, \code{selic}, and \code{cdi}.}
}
\value{
A \code{tibble}, com os valor corrigido, e os índices.
}
\description{
\code{corrigir_juros()} usa dados do IPEADATA para calcular o valor atualizado.
}
\examples{
# Exemplo, corrigindo pela selic:
# df <- correct_interest(valor = 1000, dtInicio = "01/01/2003",dtFim = Sys.Date(), indice = "selic")

}
\references{
Os índeces podem ser acessados através do acesso ao sítio do IPEA Data
\url{http://www.ipeadata.gov.br/}.
}
