% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correct_TLm.R
\name{correct_TLm}
\alias{correct_TLm}
\title{Correção de um valor pela TLm - Taxa Legal}
\usage{
correct_TLm(valor = 1e+06, dt_inicio = "02/01/2008", dt_fim = "30/08/2024")
}
\arguments{
\item{valor}{Valor a ser reajustado (default é 1000.00).}

\item{dtInicio}{Data de início da atualização.}

\item{dtFim}{Data final da atualização.}
}
\value{
A \code{tibble}, com os valor corrigido, e os índices.
}
\description{
\code{correct_TLm()} usa dados do IPEADATA para calcular o valor atualizado pela TLm.
}
\examples{
# Exemplo, corrigindo pela TLm:
# df <- correct_interest(valor = 1000000, dtInicio = "01/01/2003",dtFim = Sys.Date())

}
\references{
Os índeces podem ser acessados através do acesso ao sítio do IPEA Data
\url{http://www.ipeadata.gov.br/}.
}
