% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_dCalendario.R
\name{create_dCalendario}
\alias{create_dCalendario}
\title{Criar um Data Frame de Calendário}
\usage{
create_dCalendario(data_inicial = "2001/01/01", data_final = "2040/12/31")
}
\arguments{
\item{data_inicial}{Uma string de caracteres representando a data inicial no formato "YYYY/MM/DD". O padrão é "2001/01/01".}

\item{data_final}{Uma string de caracteres representando a data final no formato "YYYY/MM/DD". O padrão é "2040/12/31".}
}
\value{
Um tibble contendo os dados do calendário com as seguintes colunas:
\describe{
\item{data}{Data no formato "YYYY-MM-DD".}
\item{ano}{Ano extraído da data.}
\item{mes}{Mês extraído da data.}
\item{mes_nome_abr}{Nome abreviado do mês em português.}
\item{des_mes_ano}{Nome abreviado do mês e os dois últimos dígitos do ano.}
\item{mes_nome}{Nome completo do mês em português.}
\item{mes_ano}{Mês e ano concatenados com um sublinhado.}
\item{tri}{Trimestre do ano.}
\item{dia}{Dia do mês.}
\item{dia_semana}{Dia da semana como um número (1 = Domingo, 7 = Sábado).}
\item{dia_semana_nome}{Nome completo do dia da semana em português.}
\item{dia_semana_nome_abr}{Nome abreviado do dia da semana em português.}
\item{du_mes}{Dia útil do mês de acordo com o calendário "Brazil/ANBIMA".}
\item{ordem_mes_nome}{Número de ordem do grupo para cada combinação mês-ano.}
}
}
\description{
Esta função gera um data frame de calendário com informações detalhadas de datas
de uma data inicial especificada até uma data final. O calendário inclui várias
colunas relacionadas a datas, como ano, mês, dia, dia da semana e dias úteis
em um mês de acordo com o calendário "Brazil/ANBIMA".
}
\examples{
# Criar um calendário de 2021 a 2022
calendario <- create_dCalendario("2021/01/01", "2022/12/31")
head(calendario)

}
