% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cut_paste_file.R
\name{cut_paste_file}
\alias{cut_paste_file}
\title{Copy and Paste File}
\usage{
cut_paste_file(
  source_folder = NULL,
  destination_folder,
  pattern_file_name,
  new_name = NULL,
  last_modified = TRUE,
  file_path = NULL
)
}
\arguments{
\item{source_folder}{character. The path to the source folder where the function will look for the last modified file.}

\item{destination_folder}{character. The path to the destination folder where the file will be copied to.}

\item{pattern_file_name}{regex of the file name to find}

\item{new_name}{character. The new name for the file. If not provided, the function will use the original file name.}

\item{last_modified}{logical. Indicates if the function should use the last modified file in the source folder.}

\item{file_path}{character. The path to the specific file that should be copied.}
}
\value{
Nothing
}
\description{
The function is used to copy a file from a specified source folder or a specific file path to a destination folder and possibly
change the name of the file.
If a file path is provided, the function will use it, otherwise it will use the last modified file in the source folder.
}
\examples{
# copy the last modified file in the default source
# folder to the specified destination folder with the original file name
# cut_paste_file(destination_folder = "C:/path/to/destination_folder")
# copy the last modified file in the folder to the specified destination folder
# cut_paste_file(destination_folder = "C:/path/to/destination_folder", new_name = "new_name.csv")
}
