% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_files.R
\name{find_files}
\alias{find_files}
\title{Search for files in a specified path recursively}
\usage{
find_files(pattern, root_path = "C:/")
}
\arguments{
\item{pattern}{A string representation of the pattern you want to match.}

\item{root_path}{A string representation of the root path where the function should start looking for the files. Default is "C:/"}
}
\value{
a tibble with one column named "local_arquivo" that contains the file paths that match the pattern
}
\description{
Search for files in a specified path recursively
}
