% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_insert.R
\name{find_insert}
\alias{find_insert}
\title{Localize and insert values into a data frame}
\usage{
find_insert(
  base_classificar,
  nome_coluna_procurar,
  base_identificacao,
  nome_coluna_inserir
)
}
\arguments{
\item{base_classificar}{A data frame that will receive the insertion of the values.}

\item{nome_coluna_procurar}{The name of the column in the base_classificar data frame that will be searched for a value.}

\item{base_identificacao}{A reference data frame containing the values to be inserted.}

\item{nome_coluna_inserir}{The name of the column in the base_classificar data frame that will receive the insertion of the values.}
}
\value{
The base_classificar data frame with the inserted values.
}
\description{
This function is used to localize a value in a column of a data frame and insert the value of another column from a reference data frame.
}
