% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_number.R
\name{format_number}
\alias{format_number}
\title{Formatar Número}
\usage{
format_number(x, acc = 0.1)
}
\arguments{
\item{x}{O número a ser formatado.}

\item{acc}{A precisão desejada para o número formatado. O padrão é 0.1.}
}
\value{
Retorna o número formatado como uma string.
}
\description{
Esta função formata um número de acordo com o padrão Brasileiro.
}
\examples{
# format_number(123456.789)
# format_number(123456.789, acc = 0.01)
}
