% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_percent.R
\name{format_percent}
\alias{format_percent}
\title{format_percent}
\usage{
format_percent(x, acc = 0.1, scale = 100)
}
\arguments{
\item{x}{O número a ser formatado.}

\item{acc}{A precisão desejada para o número formatado. O padrão é 0.1.}

\item{scale}{A escala do numero que esta sendo passado. 1 ou 100.}
}
\value{
Retorna o número formatado como uma string.
}
\description{
Esta função formata o percentual de acordo com o padrão Brasileiro.
}
\examples{
# formatar_numero(123456.789, acc = 0.01, scale = 100)
}
