% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dollar_data.R
\name{get_dollar_data}
\alias{get_dollar_data}
\title{Fetch exchange rate data from IPEA}
\usage{
get_dollar_data()
}
\value{
A data frame with two columns: \code{Data} and \code{valor}.

A \code{tibble}, com os valor do dólar
}
\description{
The \code{dados_dolar} function fetches exchange rate data of Brazilian Real against US Dollar from IPEA. The function uses the OData API from IPEA to retrieve the data and calculates the changes in prices. The final result is a data frame with two columns: \code{Data} and \code{valor}. The \code{Data} column represents the date, and the \code{valor} column represents the exchange rate in Brazilian Real against US Dollar. The function returns a data frame and prints a message indicating the type of data returned.
}
\examples{
dolar <- get_dollar_data()

}
\references{
Os índeces podem ser acessados através do acesso ao sítio do IPEA Data
\url{http://www.ipeadata.gov.br/}.
}
