% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_interest_base.R
\name{get_interest_base}
\alias{get_interest_base}
\title{Buscar os dados do cdi ou da selic entre duas datas}
\usage{
get_interest_base(dtInicio = "01/01/2003", dtFim = Sys.Date(), indice = "cdi")
}
\arguments{
\item{dtInicio}{Data de início do período}

\item{dtFim}{Data final do período.}

\item{indice}{Indice de juros a ser utilizado (default é "cdi"). Os índices disponíveis são: cdi e selic

\code{selic}, and \code{cdi}.}
}
\value{
A \code{tibble}, com os dados dos índices.
}
\description{
\code{get_interest_base()} busca e trata os dados do cdi ou da selic do IPEADATA.
}
\examples{
# get_interest_base(dtInicio = "01/01/2003",dtFim = Sys.Date(), indice = "selic")

}
\references{
Os índeces podem ser acessados através do acesso ao sítio do IPEA Data
\url{http://www.ipeadata.gov.br/}.
}
