% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_files_information.R
\name{map_files_information}
\alias{map_files_information}
\title{Map Files Information}
\usage{
map_files_information(directory)
}
\arguments{
\item{directory}{A character string specifying the path to the directory.}
}
\value{
A tibble with information about each file in the directory and its subdirectories.
The tibble includes a column "caminho" which contains the path to each file.
}
\description{
This function lists all files in a directory and its subdirectories,
gets information about each file, converts the row names of the file
information into a column, and returns the resulting tibble.
}
\examples{
\dontrun{
file_info <- get_file_info("C:/Users")
}

}
