% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_pdfs.R
\name{merge_pdfs}
\alias{merge_pdfs}
\title{Combine PDF files in a folder and save the result}
\usage{
merge_pdfs(Caminho_pasta, pattern = "*.pdf", output_file = "joined.pdf")
}
\arguments{
\item{Caminho_pasta}{Character string specifying the path to the folder containing PDF files to be combined.}

\item{pattern}{Character string specifying the file name pattern to be used to identify PDF files in the folder.}

\item{output_file}{Character string specifying the name of the output file. Default is "joined.pdf".}
}
\value{
This function does not return a value.
}
\description{
This function identifies PDF files in a folder, combines them, and saves the result in a new file.
}
