% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_path.R
\name{prepare_path}
\alias{prepare_path}
\title{Prepare Path}
\usage{
prepare_path(path = "clipboard")
}
\arguments{
\item{path}{A string specifying the path. If "clipboard", the function will read from the clipboard. Default is "clipboard".}
}
\value{
A string with the prepared path.
}
\description{
This function prepares a path for use. It reads from the clipboard or prompts the user for input, removes quotes,
replaces backslashes with forward slashes, and then copies the result back to the clipboard.
}
\examples{
prepare_path()
}
