% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recognize_ocr.R
\name{recognize_ocr}
\alias{recognize_ocr}
\title{Recognize OCR from a PDF file and save it to a text file}
\usage{
recognize_ocr(
  filepath = "C:/Users/FLeal/Area de Trabalho/suzy/atas/2023_03_09.pdf"
)
}
\arguments{
\item{filepath}{A character string representing the file path of the PDF file to be converted to text. Defaults to
"C:/Users/FLeal/OneDrive - COMPANHIA AIX DE PARTICIPACOES/Área de Trabalho/suzy/atas/2023_03_09.pdf".}
}
\value{
This function doesn't return any value. It saves the recognized OCR from the PDF file to a text file with the same name.
}
\description{
This function takes a PDF file and converts it to text using OCR. The resulting text is then saved to a text file with the same name as the PDF file.
}
\examples{
# Exemplo, corrigindo pelo ipca:
# arquivos <- list.files("C:/Users/FLeal/Área de Trabalho/suzy/atas/", full.names = TRUE) \%>\%
# as_tibble_col("filepath")
# pmap(arquivos,reconhecer_ocr)

}
