% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reinstall_fgl_packages.R
\name{reinstall_fgl_packages}
\alias{reinstall_fgl_packages}
\title{Reinstall one of the FGL packages}
\usage{
reinstall_fgl_packages(pacote = "FGLUtils")
}
\arguments{
\item{pacote}{The name of the package to reinstall. It can be "FGLUtils", "FGLReports","FGLDerivs","pkFGLFinPe","pkAIX".
Default is "FGLUtils"}
}
\value{
None
}
\description{
Reinstall one of the FGL packages
}
\examples{
# reinstall_fgl_packages("FGLUtils")
# reinstall_fgl_packages("FGLReports")
# reinstall_fgl_packages("FGLDerivs")
# reinstall_fgl_packages("pkFGLFinPe")
# reinstall_fgl_packages("pkAIX")
}
