% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_excel.R
\name{to_excel}
\alias{to_excel}
\title{Create Excel file with a sheet for each dataframe in a list}
\usage{
to_excel(df, arquivo = "padrao", caminho = "C:/Users/FLeal/Downloads")
}
\arguments{
\item{df}{a list of dataframes}

\item{arquivo}{a character string for the name of the excel file}
}
\value{
opens the Excel file
}
\description{
Takes a list of dataframes and exports each dataframe to a separate sheet in an Excel file, with the sheet name being the name of the dataframe if it's named in the list, or "Sheet" followed by the index of the dataframe in the list if it's not named.
}
