% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_md.R
\name{to_md}
\alias{to_md}
\title{Convert a dataframe to markdown}
\usage{
to_md(your_df)
}
\arguments{
\item{df}{a dataframe}
}
\value{
print in the console the markdown format
}
\description{
Takes a dataframe and convert it to markdown format
}
