% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_net_salary.R
\name{calculate_net_salary}
\alias{calculate_net_salary}
\title{Calcular o salario liquido}
\usage{
calculate_net_salary(
  funcionario = "FGL",
  proventos,
  dependentes,
  adiantamento = 0,
  pgbl = 0
)
}
\arguments{
\item{funcionario}{Nome do funcionario}

\item{proventos}{Total da receita tributada}

\item{dependentes}{Numero de dependentes}

\item{adiantamento}{Valor recebido de adiantamento}

\item{pgbl}{Eventuais valor aportado em fundo PGBL}
}
\value{
A \code{tibble}, os dados do salario, deduções e salario liquido
}
\description{
Calcular o salario líquido de impostos
}
\examples{
# FGL <- calculate_net_salary(uncionario = "FGL", proventos = 10000, dependentes = 2)

}
