% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_insert.R
\name{find_insert}
\alias{find_insert}
\title{Localize and insert values into a data frame}
\usage{
find_insert(
  BaseClassificar = extrato2,
  NomeColunaProcurar = "historico",
  BaseIndentificacao = base_identificacao,
  NomeColunaInserir = "pkConta"
)
}
\arguments{
\item{BaseClassificar}{A data frame that will receive the insertion of the values.}

\item{NomeColunaProcurar}{The name of the column in the BaseClassificar data frame that will be searched for a value.}

\item{BaseIndentificacao}{A reference data frame containing the values to be inserted.}

\item{NomeColunaInserir}{The name of the column in the BaseClassificar data frame that will receive the insertion of the values.}
}
\value{
The BaseClassificar data frame with the inserted values.
}
\description{
This function is used to localize a value in a column of a data frame and insert the value of another column from a reference data frame.
}
